Concurrent.Capability.LocalArray

Documentation

data LocalArray a

Capability-Local variables with cache-line spacing

You can _only_ safely access these from code that is pinned to a given capability. Otherwise you might get preempted between when we check the capability # and index into the array. At which point the invariants that ensure that we can operate without a compare-and-swap even on the threaded RTS cease to hold.

newLocalArray :: MonadPrimIO m => Int -> a -> m (LocalArray a)

Calling setNumCapabilities after this is built can cause you to crash when accessing it and break invariants.

readLocalArray :: forall m a. (MonadPrimIO m, ReifiesCapability m) => LocalArray a -> Int -> m a

writeLocalArray :: (MonadPrimIO m, ReifiesCapability m) => LocalArray a -> Int -> a -> m ()

atomicModifyLocalArray :: (MonadPrimIO m, ReifiesCapability m) => LocalArray a -> Int -> (a -> (a, b)) -> m b

This is safely atomic, despite the lack of CAS.

atomicModifyLocalArray' :: (MonadPrimIO m, ReifiesCapability m) => LocalArray a -> Int -> (a -> (a, b)) -> m b

This is safely atomic, despite the lack of CAS.

modifyLocalArray :: (MonadPrimIO m, ReifiesCapability m) => LocalArray a -> Int -> (a -> a) -> m a

This is safely atomic, despite the lack of CAS.

modifyLocalArray' :: (MonadPrimIO m, ReifiesCapability m) => LocalArray a -> Int -> (a -> a) -> m a

This is safely atomic, despite the lack of CAS.

fetchModifyLocalArray :: (MonadPrimIO m, ReifiesCapability m) => LocalArray a -> Int -> (a -> a) -> m a

This is safely atomic, despite the lack of CAS.

fetchModifyLocalArray' :: (MonadPrimIO m, ReifiesCapability m) => LocalArray a -> Int -> (a -> a) -> m a

This is safely atomic, despite the lack of CAS.