Concurrent.IVar

Documentation

data IVar s a

An IVar is a promise that can be fulfilled multiple times, but must be always be fulfilled with the same answer.

Constructors

IVar !(MVar a) a

newEmptyIVar :: MonadPar d i s m => m (IVar s a)

newIVar :: a -> IVar s a

Build a filled IVar out of a known value

readIVar :: IVar s a -> a

writeIVar :: (MonadPar d i s m, Eq a) => IVar s a -> a -> m ()

unsafeWriteIVar :: MonadPar d i s m => IVar s a -> a -> m ()

Like writeIVar but assumes (without checking) that the write is always consistent.