Concurrent.Primitive.Array

Array Primitives

sizeofArray :: Array a -> Int

sizeofMutableArray :: MutableArray s a -> Int

casArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> a -> a -> m (Int, a)

atomicModifyArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> (a, b)) -> m b

atomicModifyArray' :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> (a, b)) -> m b

modifyArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position. Return the new result

modifyArray m i f = atomicModifyArray m i $ a -> let b = f a in (b, b)

modifyArray' :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position strictly. Return the new result.

Can this be smarter? e.g. start it off already as a blackhole we appear to be evaluating, putting frames on the stack, etc. That would avoid anybody ever getting and seeing the unevaluated closure.

fetchModifyArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position. Return the old result.

fetchModifyArray' :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position strictly. Return the old result.

Capability-local array manipulation primitives

localAtomicModifyArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> (a, b)) -> m b

Modify the contents of a position in an array in a manner that at least can't be preempted another thread in the current capability.

localAtomicModifyArray' :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> (a, b)) -> m b

Modify the contents of a position in an array strictly in a manner that at least can't be preempted another thread in the current capability.

localModifyArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position. Return the new result.

Logically,

localModifyArray m i f = localAtomicModifyArray m i $ a -> let b = f a in (b, b)

but it is a bit more efficient.

localModifyArray' :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position strictly. Return the new result.

Can this be smarter? e.g. start it off already as a blackhole we appear to be evaluating, putting frames on the stack, etc. That would avoid anybody ever getting and seeing the unevaluated closure.

localFetchModifyArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position. Return the old result.

localFetchModifyArray' :: PrimMonad m => MutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position strictly. Return the old result.

ByteArray Primitives

sizeOfByteArray :: ByteArray -> Int

sizeOfMutableByteArray :: MutableByteArray s -> Int

casIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> Int -> Int -> m Int

atomicReadIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> m Int

atomicWriteIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> Int -> m ()

fetchAddIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> Int -> m Int

fetchSubIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> Int -> m Int

fetchAndIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> Int -> m Int

fetchNandIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> Int -> m Int

fetchOrIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> Int -> m Int

fetchXorIntArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> Int -> m Int

Prefetching

prefetchByteArray0 :: PrimMonad m => ByteArray -> Int -> m ()

prefetchByteArray1 :: PrimMonad m => ByteArray -> Int -> m ()

prefetchByteArray2 :: PrimMonad m => ByteArray -> Int -> m ()

prefetchByteArray3 :: PrimMonad m => ByteArray -> Int -> m ()

prefetchMutableByteArray0 :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> m ()

prefetchMutableByteArray1 :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> m ()

prefetchMutableByteArray2 :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> m ()

prefetchMutableByteArray3 :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> m ()

prefetchValue0 :: PrimMonad m => a -> m ()

prefetchValue1 :: PrimMonad m => a -> m ()

prefetchValue2 :: PrimMonad m => a -> m ()

prefetchValue3 :: PrimMonad m => a -> m ()