Concurrent.Primitive.SmallArray

Description

Small primitive boxed arrays

Documentation

data SmallArray a

Boxed arrays

Constructors

SmallArray (SmallArray# a)

data SmallMutableArray s a

Mutable boxed arrays associated with a primitive state token.

Constructors

SmallMutableArray (SmallMutableArray# s a)

newSmallArray :: PrimMonad m => Int -> a -> m (SmallMutableArray (PrimState m) a)

Create a new mutable array of the specified size and initialise all elements with the given value.

readSmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> m a

Read a value from the array at the given index.

writeSmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> a -> m ()

Write a value to the array at the given index.

indexSmallArray :: SmallArray a -> Int -> a

Read a value from the immutable array at the given index.

indexSmallArrayM :: Monad m => SmallArray a -> Int -> m a

Monadically read a value from the immutable array at the given index. This allows us to be strict in the array while remaining lazy in the read element which is very useful for collective operations. Suppose we want to copy an array. We could do something like this:

copy marr arr ... = do ...
                       writeSmallArray marr i (indexSmallArray arr i) ...
                       ...

But since primitive arrays are lazy, the calls to indexSmallArray will not be evaluated. Rather, marr will be filled with thunks each of which would retain a reference to arr. This is definitely not what we want!

With indexSmallArrayM, we can instead write

copy marr arr ... = do ...
                       x <- indexSmallArrayM arr i
                       writeSmallArray marr i x
                       ...

Now, indexing is executed immediately although the returned element is still not evaluated.

unsafeFreezeSmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> m (SmallArray a)

Convert a mutable array to an immutable one without copying. The array should not be modified after the conversion.

unsafeThawSmallArray :: PrimMonad m => SmallArray a -> m (SmallMutableArray (PrimState m) a)

Convert an immutable array to an mutable one without copying. The immutable array should not be used after the conversion.

sameSmallMutableArray :: SmallMutableArray s a -> SmallMutableArray s a -> Bool

Check whether the two arrays refer to the same memory block.

copySmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
destination array

-> Int
offset into destination array

-> SmallArray a
source array

-> Int
offset into source array

-> Int
number of elements to copy

-> m ()

Copy a slice of an immutable array to a mutable array.

copySmallMutableArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
destination array

-> Int
offset into destination array

-> SmallMutableArray (PrimState m) a
source array

-> Int
offset into source array

-> Int
number of elements to copy

-> m ()

Copy a slice of a mutable array to another array. The two arrays may not be the same.

cloneSmallArray

Arguments

:: SmallArray a
source array

-> Int
offset into destination array

-> Int
number of elements to copy

-> SmallArray a

Return a newly allocated SmallArray with the specified subrange of the provided SmallArray. The provided SmallArray should contain the full subrange specified by the two Ints, but this is not checked.

cloneSmallMutableArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
source array

-> Int
offset into destination array

-> Int
number of elements to copy

-> m (SmallMutableArray (PrimState m) a)

Return a newly allocated SmallMutableArray. with the specified subrange of the provided SmallMutableArray. The provided SmallMutableArray should contain the full subrange specified by the two Ints, but this is not checked.

casSmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> a -> a -> m (Int, a)

Perform an unsafe, machine-level atomic compare and swap on an element within an array.

sizeOfSmallArray :: SmallArray a -> Int

sizeOfSmallMutableArray :: SmallMutableArray s a -> Int

Atomic modification

atomicModifySmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> (a, b)) -> m b

atomicModifySmallArray' :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> (a, b)) -> m b

modifySmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position. Return the new result

modifySmallArray m i f = atomicModifySmallArray m i $ a -> let b = f a in (b, b)

modifySmallArray' :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position strictly. Return the new result.

Can this be smarter? e.g. start it off already as a blackhole we appear to be evaluating, putting frames on the stack, etc. That would avoid anybody ever getting and seeing the unevaluated closure.

fetchModifySmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position. Return the old result.

fetchModifySmallArray' :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position strictly. Return the old result.

localAtomicModifySmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> (a, b)) -> m b

Modify the contents of a position in an array in a manner that at least can't be preempted another thread in the current capability.

localAtomicModifySmallArray' :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> (a, b)) -> m b

Modify the contents of a position in an array strictly in a manner that at least can't be preempted another thread in the current capability.

localModifySmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position. Return the new result.

Logically,

localModifySmallArray m i f = localAtomicModifySmallArray m i $ a -> let b = f a in (b, b)

but it is a bit more efficient.

localModifySmallArray' :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position strictly. Return the new result.

Can this be smarter? e.g. start it off already as a blackhole we appear to be evaluating, putting frames on the stack, etc. That would avoid anybody ever getting and seeing the unevaluated closure.

localFetchModifySmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position. Return the old result.

localFetchModifySmallArray' :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> (a -> a) -> m a

Modify the contents of an array at a given position strictly. Return the old result.