Concurrent.Promise.Unsafe

Documentation

data Promise s a

An Promise can be fulfilled multiple times, but must be always be fulfilled with the same answer.

Constructors

Promise !(MVar a) a

unsafeWritePromise :: MonadPar d i s m => Promise s a -> a -> m ()

Like writePromise but assumes (without checking) that the write is always consistent.