Concurrent.Promise

Documentation

data Promise s a

An Promise can be fulfilled multiple times, but must be always be fulfilled with the same answer.

Constructors

Promise !(MVar a) a

newEmptyPromise :: MonadPar d i s m => m (Promise s a)

newPromise :: a -> Promise s a

Build a filled Promise out of a known value

readPromise :: Promise s a -> a

writePromise :: (MonadPar d i s m, Eq a) => Promise s a -> a -> m ()

unsafeWritePromise :: MonadPar d i s m => Promise s a -> a -> m ()

Like writePromise but assumes (without checking) that the write is always consistent.